Store a historical registry when a product was sold

+++ src/main/java/io/edrb/stockservice/model/ProductsSoldHistorical.java
@@ -29,6 +29,6 @@ public class ProductsSoldHistorical {
     private String productId;
 
     @NotNull
-    private Long itemsSold;
+    private Integer itemsSold;
 
 }
